Create version 1.0.0 for public release (#2)
This commit contains all of the features and tests for the initial
public release of Structured Elements. It has been bundled into a single
commit on 'main' for the sake of tidiness. We intend to use pull
requests for all changes, so it seems reasonable to start with one.

For a more detailed breakdown of the features included in this version,
take a look at `README.md` or browse the code from this commit. The main
index file from `src/index.ts` might be a helpful place to start.